+++ cmd/server.go
@@ -3,156 +3,24 @@ package main
 import (
 	"context"
 	"encoding/base64"
-	"html/template"
-	"io/fs"
-	"io/ioutil"
 	"log"
 	"net/http"
-	"net/url"
-	"os"
-	"path"
-	"path/filepath"
-	"regexp"
 
 	"github.com/fsnotify/fsnotify"
 	"github.com/gorilla/websocket"
-	"github.com/kelseyhightower/envconfig"
 )
 
-var config configuration
-
-func init() {
-	err := envconfig.Process("CUEWATCH", &config)
-	if err != nil {
-		log.Fatal(err.Error())
-	}
-}
-
-type configuration struct {
-	PlantUMLServerAddress string `envconfig:"PLANTML_SERVER" default:"http://localhost:8080"`
-	ListenAddress         string `envconfig:"ADDR" default:":9090"`
-	PollingDir            string `envconfig:"POLLING_DIR" default:"./"`
-	RecursivePoll         bool   `envconfig:"RECURSIVE" default:"true"`
-}
-
 type diagram struct {
-	image    []byte
+	svgImage []byte
 	cue      []byte
 	cueErr   []byte
 	plantuml []byte
 }
 
-func main() {
-	upgrader := websocket.Upgrader{
-		ReadBufferSize:  1024,
-		WriteBufferSize: 1024,
-	}
-
-	watcher, err := fsnotify.NewWatcher()
-	if err != nil {
-		log.Fatal("NewWatcher failed: ", err)
-	}
-
-	err = watcher.Add(config.PollingDir)
-	if err != nil {
-		log.Fatal("Add failed:", err)
-	}
-	cwd, err := os.Getwd()
-	if err != nil {
-		log.Fatal(err)
-	}
-	if config.RecursivePoll {
-		tmpl, err := template.New("index").Parse(index)
-		if err != nil {
-			log.Fatal(err)
-		}
-		err = filepath.WalkDir(config.PollingDir, func(p string, d fs.DirEntry, err error) error {
-			if err != nil {
-				return err
-			}
-			if !d.IsDir() {
-				return nil
-			}
-			if isGit, err := regexp.MatchString(".*/.git/?.*", p); isGit || err != nil {
-				return nil
-			}
-			watcher, err := fsnotify.NewWatcher()
-			if err != nil {
-				log.Fatal("NewWatcher failed: ", err)
-			}
-
-			err = watcher.Add(filepath.Clean(filepath.Join(cwd, p)))
-			if err != nil {
-				return err
-			}
-			cleanP := path.Clean(path.Join("/", p))
-			log.Printf("Registering %v", cleanP)
-			me, err := url.Parse("http://" + config.ListenAddress)
-			if err != nil {
-				return err
-			}
-			if me.Hostname() == "" {
-				me.Host = "localhost" + me.Host
-				me.Path = path.Join(cleanP, "/connws/")
-			}
-			files, err := ioutil.ReadDir(p)
-			if err != nil {
-				return err
-			}
-			type dirEntry struct {
-				Path string
-				Name string
-			}
-			dirEntries := make([]dirEntry, 0)
-			for _, f := range files {
-				if f.IsDir() {
-					dirEntries = append(dirEntries, dirEntry{
-						Name: f.Name(),
-						Path: path.Clean(path.Join(cleanP, f.Name())),
-					})
-				}
-			}
-			http.HandleFunc(cleanP, func(w http.ResponseWriter, r *http.Request) {
-				err := tmpl.Execute(w, struct {
-					URL        *url.URL
-					Name       string
-					DirEntries []dirEntry
-				}{
-					DirEntries: dirEntries,
-					Name:       d.Name(),
-					URL:        me,
-				})
-				if err != nil {
-					log.Fatal(err)
-				}
-
-			})
-			http.HandleFunc(path.Join(cleanP, "connws/"), (&client{
-				upgrader: upgrader,
-				watcher: watchedDir{
-					Watcher: watcher,
-					path:    p,
-				},
-				watch: watch,
-			}).ConnWs)
-			return nil
-		})
-		if err != nil {
-			log.Fatal(err)
-		}
-	} else {
-		panic("todo")
-	}
-	err = http.ListenAndServe(config.ListenAddress, nil)
-	if err != nil {
-		log.Fatal("ListenAndServe: ", err)
-	}
-}
-
-type client struct {
+type connexionHandler struct {
 	upgrader websocket.Upgrader
 	watcher  watchedDir
-	watch    func(ctx context.Context, watcher watchedDir, C chan diagram)
+	watch    func(ctx context.Context, watcher watchedDir, C chan diagram, errC chan error)
 }
 
 type watchedDir struct {
@@ -160,8 +28,9 @@ type watchedDir struct {
 	*fsnotify.Watcher
 }
 
-func (c *client) ConnWs(w http.ResponseWriter, r *http.Request) {
+func (c *connexionHandler) ConnWs(w http.ResponseWriter, r *http.Request) {
 	C := make(chan diagram, 5)
+	errC := make(chan error, 5)
 	log.Printf("New Connection with %v", C)
 	defer func() {
 		// drain and close
@@ -170,7 +39,7 @@ func (c *client) ConnWs(w http.ResponseWriter, r *http.Request) {
 		}
 		close(C)
 	}()
-	go c.watch(r.Context(), c.watcher, C)
+	go c.watch(r.Context(), c.watcher, C, errC)
 	ws, err := c.upgrader.Upgrade(w, r, nil)
 	if _, ok := err.(websocket.HandshakeError); ok {
 		http.Error(w, "Not a websocket handshake", 400)
@@ -190,7 +59,7 @@ func (c *client) ConnWs(w http.ResponseWriter, r *http.Request) {
 		}
 		select {
 		case diagram := <-C:
-			str := base64.StdEncoding.EncodeToString(diagram.image)
+			str := base64.StdEncoding.EncodeToString(diagram.svgImage)
 			res["image"] = str
 			res["cue"] = diagram.cueErr
 			res["plantuml"] = diagram.plantuml
@@ -199,6 +68,13 @@ func (c *client) ConnWs(w http.ResponseWriter, r *http.Request) {
 				log.Println(err)
 			}
 			log.Println("new picture sent")
+		case err := <-errC:
+			log.Println("error receiver:", err)
+			res["error"] = []byte(err.Error())
+			if err = ws.WriteJSON(&res); err != nil {
+				log.Println(err)
+			}
+			delete(res, "error")
 		case <-r.Context().Done():
 			return
 		}
